% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_diabetes_gcomputation.R
\name{analyse_diabetes_gcomputation}
\alias{analyse_diabetes_gcomputation}
\title{Analyse data set with G-computation}
\usage{
analyse_diabetes_gcomputation()
}
\value{
A function that, when called with \code{condition} and \code{dat}, returns a list with:
\itemize{
\item \code{coef}      estimated difference in mean change in HbA1c between treatment groups
\item \code{se}        standard error for coef
\item \code{p}         p-value for coef
\item \code{ci_lower}  lower bound of 95\% confidence interval for coef
\item \code{ci_upper}  upper bound of 95\% confidence interval for coef
}
}
\description{
Analyse data set with G-computation
}
\details{
This function implements G-computation to estimate the
average treatment effect (ATE) on the \strong{change in HbA1c} under a hypothetical scenario
where rescue medication is not available, regardless of treatment discontinuation.

Two scenarios are considered depending on \code{setup}:
\itemize{
\item \code{setup = 1}: the effect of rescue goes on top of the effect of the assigned treatment
\item \code{setup = 0}: once rescue starts, treatment is deterministically recoded to control
}

The estimand is:
\itemize{
\item \strong{Endpoint}: Change in HbA1c at final visit k
\item \strong{Treatments}: Experimental vs. control
\item \strong{Summary measure}: Difference in means
\item \strong{Population}: People with type 2 diabetes
\item \strong{Intercurrent events}: Use of rescue medication and treatment discontinuation
\item \strong{Intercurrent event strategy}:
\itemize{
\item \emph{Hypothetical strategy}: Assume rescue medication was not available (i.e., rescue = 0 always)
\item \emph{Treatment policy strategy}: Ignore treatment discontinuation (i.e., continue treatment in counterfactual)
}
}

The null hypothesis of no treatment effect is rejected if the 95\% CI does not include 0.
}
\examples{
\donttest{
Design <- diabetes_scenario()[1, ] |>
  diabetes_scenario_set_truevalues()

dat <- generate_diabetes(Design)

analyse_diabetes_gcomputation()(Design, dat)
}
}
