% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_MACE.R
\name{analyse_mace_ipw_nocov}
\alias{analyse_mace_ipw_nocov}
\title{Analysis function with parameter extraction : ipw model without covariates}
\usage{
analyse_mace_ipw_nocov(condition, dat, fixed_objects = NULL)
}
\arguments{
\item{condition}{SimDesign object}

\item{dat}{SimDesign object}

\item{fixed_objects}{SimDesign object}
}
\value{
named list of results
coef = estimated treatment effect
se = standard error of the coef
HR = hazard ratio
CI.l CI.u = lower and upper CI of HR
}
\description{
Analysis function with parameter extraction : ipw model without covariates
}
\examples{
internal use by SimDesign
}
