% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_minimal_example.R
\name{analyse_minimal_example_lm}
\alias{analyse_minimal_example_lm}
\title{Analyse dataset from minimal example scenario with linear regression}
\usage{
analyse_minimal_example_lm(ci_level = 0.95)
}
\arguments{
\item{ci_level}{the confidence level for the CIs (defaults to 0.95)}
}
\value{
an analyse function that returns a list with the elements
\itemize{
\item \code{p} the p-value of the F-test
\item \code{coef} the point estimate
\item \code{ci_lower} the lower CI limit
\item \code{ci_upper} the upper CI limit
}
}
\description{
Analyse dataset from minimal example scenario with linear regression
}
\examples{
Design <- assumptions_minimal_example() |>
  true_summary_statistics_minimal_example()

condition <- Design[1, ]

dat <- generate_minimal_example(condition)

my_analyse_lm <- analyse_minimal_example_lm(ci_level=0.9)
my_analyse_lm(condition, dat)
}
