% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_minimal_example.R
\name{analyse_minimal_example_t}
\alias{analyse_minimal_example_t}
\title{Analyse dataset from minimal example scenario with t-test}
\usage{
analyse_minimal_example_t()
}
\value{
an analyse function that returns a list with the elements
\itemize{
\item \code{p} the p-value
}
}
\description{
Analyse dataset from minimal example scenario with t-test
}
\examples{
Design <- assumptions_minimal_example() |>
  true_summary_statistics_minimal_example()

condition <- Design[1, ]

dat <- generate_minimal_example(condition)

my_analyse_t  <- analyse_minimal_example_t()
my_analyse_t(condition, dat)
}
