% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_oncology_cens.R
\name{analyse_oncology_cens}
\alias{analyse_oncology_cens}
\title{Create Analyse Functions for}
\usage{
analyse_oncology_cens(X)
}
\arguments{
\item{X}{input can be used to pass parameters to the analyse function}
}
\value{
an analyse function that can be used in runSimulation
}
\description{
Create Analyse Functions for
}
\examples{
setting <- oncology_scenario()[1, ]
dat <- generate_oncology(setting)
analyse_oncology_cens()(setting, dat)
}
