% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_oncology_iptw2.R
\name{analyse_oncology_ipw2}
\alias{analyse_oncology_ipw2}
\title{Create Analyse Functions for}
\usage{
analyse_oncology_ipw2(
  use_censoring_IPW = FALSE,
  trunc_weights = 5,
  requ_n_cens = 5
)
}
\arguments{
\item{use_censoring_IPW}{logical, default is FALSE}

\item{trunc_weights}{numeric}

\item{requ_n_cens}{numeric}
}
\value{
an analyse function that can be used in runSimulation
}
\description{
Create Analyse Functions for
}
\examples{
setting <- oncology_scenario()[1, ]
dat <- generate_oncology(setting)
analyse_oncology_ipw2()(setting, dat)
}
