% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_oncology_itt.R
\name{analyse_oncology_itt}
\alias{analyse_oncology_itt}
\title{Create Analyse Functions for ITT analysis}
\usage{
analyse_oncology_itt(X)
}
\arguments{
\item{X}{input can be used to pass parameters to the analyse function}
}
\value{
an analyse function that can be used in runSimulation
}
\description{
Create Analyse Functions for ITT analysis
}
\examples{
setting <- oncology_scenario()[1, ]
dat <- generate_oncology(setting)
analyse_oncology_itt()(setting, dat)
}
