% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_oncology_TSE.R
\name{analyse_oncology_TSE}
\alias{analyse_oncology_TSE}
\title{Create Analyse Functions for}
\usage{
analyse_oncology_TSE(recensor = TRUE, alpha = 0.05, B = 100)
}
\arguments{
\item{X}{input can be used to pass parameters to the analyse function}
}
\value{
an analyse function that can be used in runSimulation
}
\description{
Create Analyse Functions for
}
\examples{
\donttest{
setting <- oncology_scenario()[1, ]
dat <- generate_oncology(setting)
analyse_oncology_TSE()(setting, dat)
}
}
