% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_vaccine_ivreg2.R
\name{analyse_vaccine_ivreg2}
\alias{analyse_vaccine_ivreg2}
\title{Analyse dataset from vaccine scenario using instrumental variable regression}
\usage{
analyse_vaccine_ivreg2(
  ci_level = 0.95,
  VE_margin = 0.3,
  V_unobserved = FALSE,
  W_unobserved = FALSE
)
}
\arguments{
\item{ci_level}{the confidence level for the CIs (defaults to 0.95)}

\item{VE_margin}{vaccine efficacy margin for the super-superiority test}

\item{V_unobserved}{consider covariate V unobserved (don't use it in analysis)}

\item{W_unobserved}{consider covariate W unobserved (don't use it in analysis)}
}
\value{
an analyse function that returns a list with the elements
\itemize{
\item \code{p_model} p-value of the super-superiority test, model based vcov
\item \code{VE} the point estimate for the vaccine efficacy, model based vcov
\item \code{VE_lower} the lower CI limit for vaccine efficacy, model based vcov
\item \code{VE_upper} the upper CI limit for vaccine efficacy, model based vcov
\item \code{p} the p-value of the super-superiority test, robust vcov
\item \code{VE_sandwich} the point estimate for the vaccine efficacy, robust vcov
\item \code{VE_lower_sandwich} the lower CI limit for vaccine efficacy, robust vcov
\item \code{VE_upper_sandwich} the upper CI limit for vaccine efficacy, robust vcov
\item \code{N_pat} number of patients in the dataset
\item \code{N_evt} number of observed events in the dataset
}
}
\description{
Analyse dataset from vaccine scenario using instrumental variable regression
}
\details{
This implementation uses a different first stage than \code{analyse_vaccine_ivreg}.
}
\examples{
Design <- vaccine_scenario()

condition <- Design[3, ]
condition$p_W <- 0.3

condition <- condition  |>
  vaccine_scenario_set_beta_A1_relative() |>
  vaccine_scenario_set_gamma_0() |>
  vaccine_scenario_set_true_eff() |>
  vaccine_scenario_set_samplesize()

dat <- generate_vaccine(condition)
my_analyse  <- analyse_vaccine_ivreg (ci_level=0.95)
my_analyse2 <- analyse_vaccine_ivreg2(ci_level=0.95)

res1 <- my_analyse(conditon, dat)
res2 <- my_analyse2(conditon, dat)

res1
res2

lapply(names(res1), \(i){res1[[i]]-res2[[i]]})
}
