% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_vaccine_pp.R
\name{analyse_vaccine_pp}
\alias{analyse_vaccine_pp}
\title{Analyse dataset from vaccine scenario, per-protocol analysis}
\usage{
analyse_vaccine_pp(
  ci_level = 0.95,
  VE_margin = 0.3,
  V_unobserved = FALSE,
  W_unobserved = FALSE
)
}
\arguments{
\item{ci_level}{the confidence level for the CIs (defaults to 0.95)}

\item{VE_margin}{vaccine efficacy margin for the super-superiority test}

\item{V_unobserved}{consider covariate V unobserved (don't use it in analysis)}

\item{W_unobserved}{consider covariate W unobserved (don't use it in analysis)}
}
\value{
an analyse function that returns a list with the elements
\itemize{
\item \code{p} the p-value of the super-superiority test
\item \code{VE} the point estimate for the vaccine efficacy
\item \code{VE_lower} the lower CI limit
\item \code{VE_upper} the upper CI limit
}
}
\description{
Estimates and tests are based on marginal estimates from a Poisson regression
adjusted for \code{V} and \code{W}.
}
\examples{
Design <- vaccine_scenario() |>
  vaccine_scenario_set_beta_A1_relative() |>
  vaccine_scenario_set_gamma_0() |>
  vaccine_scenario_set_true_eff() |>
  vaccine_scenario_set_samplesize()
Design

dat <- generate_vaccine(Design[1,])
my_analyse <- analyse_vaccine_pp(ci_level=0.95)
my_analyse(Design[1, ], dat)
}
