% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_vaccine_pp2.R
\name{analyse_vaccine_pp2}
\alias{analyse_vaccine_pp2}
\title{Analyse dataset from vaccine scenario, per-protocol analysis}
\usage{
analyse_vaccine_pp2(ci_level = 0.95, VE_margin = 0.3)
}
\arguments{
\item{ci_level}{the confidence level for the CIs (defaults to 0.95)}

\item{VE_margin}{vaccine efficacy margin for the super-superiority test}
}
\value{
an analyse function that returns a list with the elements
\itemize{
\item \code{VE} the point estimate for the vaccine efficacy
\item \code{VE_lower} the lower CI limit for vaccine efficacy
\item \code{VE_upper} the upper CI limit for vaccine efficacy
\item \code{p} the p-value of the exact super-superiority test
\item \code{N_pat} number of patients in the dataset
\item \code{N_evt} number of observed events in the dataset
}
}
\description{
Estimates and tests are based on an exact conditional test. As described in
Nauta et. al.
}
\examples{
Design <- vaccine_scenario() |>
  vaccine_scenario_set_beta_A1_relative() |>
  vaccine_scenario_set_gamma_0() |>
  vaccine_scenario_set_true_eff() |>
  vaccine_scenario_set_samplesize()
Design

dat <- generate_vaccine(Design[1,])
my_analyse <- analyse_vaccine_pp2(ci_level=0.95)
my_analyse(Design[1, ], dat)
}
