% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_diabetes_rescue.R
\name{assumptions_diabetes_rescue}
\alias{assumptions_diabetes_rescue}
\title{Create an empty assumptions data.frame for generate_diabetes_rescue}
\usage{
assumptions_diabetes_rescue(print = interactive())
}
\arguments{
\item{print}{print code to generate parameter set?}
}
\value{
For assumptions_diabetes_rescue: a design tibble with default values invisibly
}
\description{
Create an empty assumptions data.frame for generate_diabetes_rescue
}
\details{
assumptions_diabetes_rescue generates a default design \code{data.frame}
for use with generate_diabetes_rescue If print is \code{TRUE} code to produce
the template is also printed for copying, pasting and editing by the user.
(This is the default when run in an interactive session.)
}
\section{Functions}{
\itemize{
\item \code{assumptions_diabetes_rescue()}: generate default design tibble

}}
\examples{
Design <- assumptions_diabetes_rescue()[1, ]
Design
}
