% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_MACE.R
\name{censor_buffer_window}
\alias{censor_buffer_window}
\title{Function that censor MACE events after discontinuation + buffer window
t_mace = T_disc + buffer is censored at one year}
\usage{
censor_buffer_window(data, buffer)
}
\arguments{
\item{data}{data in wide format, output of generate_mace function}

\item{buffer}{real non-negative value, duration of the buffer window}
}
\value{
data in wide format, with MACE events censored after the buffer window
}
\description{
Function that censor MACE events after discontinuation + buffer window
t_mace = T_disc + buffer is censored at one year
}
\examples{
Design <- mace_scenario()
toy_data <- generate_mace(Design[1,])
buffer <- as.numeric( Design[1,"true_window"])
toy_data_trunc <- censor_buffer_window(toy_data,buffer)
}
