% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_MACE.R
\name{cox_model_ipw}
\alias{cox_model_ipw}
\title{Cox model with IPW weights with covariates}
\usage{
cox_model_ipw(data)
}
\arguments{
\item{data}{simulated data with long format, output of data_process_start_stop function}
}
\value{
a coxph fit
}
\description{
Cox model with IPW weights with covariates
}
\examples{
toy_data <- generate_mace(Design[1,])
toy_data_trunc <- censor_buffer_window(toy_data,buffer)
toy_data_trunc_long <- data_process_start_stop(toy_data_trunc)
cox_model_ipw(toy_data_trunc_long)
}
