% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_MACE.R
\name{cox_model_nocov}
\alias{cox_model_nocov}
\title{Cox model without covariates}
\usage{
cox_model_nocov(data)
}
\arguments{
\item{data}{data in wide format, output of censor_buffer window}
}
\value{
a coxph fit
}
\description{
Cox model without covariates
}
\examples{
Design <- mace_scenario()
toy_data <- generate_mace(Design[1,])
buffer <- as.numeric( Design[1,"true_window"])
toy_data_trunc <- censor_buffer_window(toy_data,buffer)
cox_model_nocov(toy_data_trunc)
}
