% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_MACE.R
\name{data_process_start_stop}
\alias{data_process_start_stop}
\title{Function that transforms data set with t_mace,event_mace, with one row per patients
into a data set with t_mace_start, t_mace_stop, event_mace, with discontinuation as
a time varying covariate}
\usage{
data_process_start_stop(data)
}
\arguments{
\item{data}{simulated data in wide format, output of generate_mace and censor_buffer_window}
}
\value{
data in long format for IPW analysis
}
\description{
Function that transforms data set with t_mace,event_mace, with one row per patients
into a data set with t_mace_start, t_mace_stop, event_mace, with discontinuation as
a time varying covariate
}
\examples{
Design <- mace_scenario()
toy_data <- generate_mace(Design[1,])
buffer <- as.numeric( Design[1,"true_window"])
toy_data_trunc <- censor_buffer_window(toy_data,buffer)
toy_data_trunc_long <- data_process_start_stop(toy_data_trunc)
}
