% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_diabetes.R
\name{diabetes_scenario}
\alias{diabetes_scenario}
\title{Create an empty assumptions data.frame for generate_diabetes}
\usage{
diabetes_scenario(print = interactive())
}
\arguments{
\item{print}{print code to generate parameter set?}
}
\value{
For diabetes_scenario: a design tibble with default values invisibly
}
\description{
Create an empty assumptions data.frame for generate_diabetes
}
\details{
diabetes_scenario generates a default design \code{data.frame}
for use with generate_diabetes If print is \code{TRUE} code to produce
the template is also printed for copying, pasting and editing by the user.
(This is the default when run in an interactive session.)
}
\section{Functions}{
\itemize{
\item \code{diabetes_scenario()}: generate default design tibble

}}
\examples{
Design <- diabetes_scenario()[1, ]
Design
}
