% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MACE.R
\name{generate_all_tte2}
\alias{generate_all_tte2}
\title{For internal use only}
\usage{
generate_all_tte2(cov_df, condition = condition, no_withdraw = FALSE)
}
\arguments{
\item{cov_df}{dataset of covariates generated inside generate_mace function}

\item{condition}{passed on from within generate_mace call}

\item{no_withdraw}{if true sets prob_withdraw to 0}
}
\value{
a simulated Dataset
}
\description{
For internal use only
}
