% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_diabetes.R
\name{generate_diabetes}
\alias{generate_diabetes}
\alias{diabetes_scenario_set_truevalues}
\title{Generate Dataset that simulates a diabetes trial in which rescue medication
can be introduced}
\usage{
generate_diabetes(condition, fixed_objects = NULL)

diabetes_scenario_set_truevalues(Design)
}
\arguments{
\item{condition}{condition row of Design dataset}

\item{fixed_objects}{fixed objects of Design dataset}

\item{Design}{Design data.frame for diabetes scenarios, e.g. created with diabetes_scenario()}
}
\value{
For generate_diabetes: A data set with n rows and the columns id, trt
(1=treatment, 0=control), age, y0, y1, ..., yk
(the repeated measurements of the outcome), R1, ..., Rk-1 (indicators for
rescue medication at each visit except baseline and the last visit),
and rescue_start (the visit at which rescue medication was started,
NA if no rescue medication was received).

For diabetes_scenario_set_truevalues: a design tibble with true values for the treatment effect at final visit and the sample size needed to achieve 80\% power for a two-sided test at alpha = 0.05
}
\description{
Generate Dataset that simulates a diabetes trial in which rescue medication
can be introduced

Calculate true summary statistics for scenarios with delayed treatment effect
}
\details{
Condition has to contain the following columns:
\itemize{
\item k
\item mean_age
\item sd_age
\item b_age
\item mean_bl
\item sd_bl
\item rho
\item delta
\item lambda
\item delta_resc
\item lambda_resc
\item resc_0
\item resc_y
\item resc_age
\item miss
}
}
\section{Functions}{
\itemize{
\item \code{generate_diabetes()}: simulates a data set with n rows.

\item \code{diabetes_scenario_set_truevalues()}: calculate true summary statistics for ...

}}
\examples{
Design <- diabetes_scenario()
generate_diabetes(Design[1, ])
diabetes_scenario_set_truevalues(diabetes_scenario())
}
