% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_diabetes_rescue.R
\name{generate_diabetes_rescue}
\alias{generate_diabetes_rescue}
\alias{assumptions_diabetes_rescue}
\alias{true_summary_statistics_diabetes_rescue}
\title{Generate Dataset ...}
\usage{
generate_diabetes_rescue(condition, fixed_objects = NULL)

assumptions_diabetes_rescue(print = interactive())

true_summary_statistics_diabetes_rescue(
  Design,
  cutoff_stats = 10,
  fixed_objects = NULL
)
}
\arguments{
\item{condition}{condition row of Design dataset}

\item{fixed_objects}{fixed objects not used for now}

\item{print}{print code to generate parameter set?}

\item{Design}{Design data.frame for x}

\item{cutoff_stats}{Cutoff time for rmst and average hazard ratios}
}
\value{
For generate_diabetes_rescue: A data set with the columns id, trt
(1=treatment, 0=control), evt (event, currently TRUE for all observations)

For assumptions_diabetes_rescue: a design tibble with default values invisibly

For true_summary_statistics_x: the design data.frame
passed as argument with the additional columns:
\itemize{
\item \code{rmst_trt} rmst in the treatment group
\item \code{median_surv_trt} median survival in the treatment group
\item \code{rmst_ctrl} rmst in the control group
\item \code{median_surv_ctrl} median survial in the control group
\item \code{gAHR} geometric average hazard ratio
\item \code{AHR} average hazard ratio
}
}
\description{
Generate Dataset ...

Create an empty assumptions data.frame for generate_diabetes_rescue

Calculate true summary statistics for scenarios with delayed treatment effect
}
\details{
Condition has to contain the following columns:
\itemize{
\item eff the effect size
\item 
\item ...
}

assumptions_diabetes_rescue generates a default design \code{data.frame}
for use with generate_diabetes_rescue If print is \code{TRUE} code to produce
the template is also printed for copying, pasting and editing by the user.
(This is the default when run in an interactive session.)
}
\section{Functions}{
\itemize{
\item \code{generate_diabetes_rescue()}: simulates a dataset with ...

\item \code{assumptions_diabetes_rescue()}: generate default design tibble

\item \code{true_summary_statistics_diabetes_rescue()}: calculate true summary statistics for ...

}}
\examples{
Design <- assumptions_diabetes_rescue()
Design
}
