% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_minimal_example.R
\name{assumptions_minimal_example}
\alias{assumptions_minimal_example}
\alias{true_summary_statistics_minimal_example}
\alias{generate_minimal_example}
\title{Create an empty assumtions tibble for generate_minimal_example}
\usage{
assumptions_minimal_example(print = interactive())

true_summary_statistics_minimal_example(Design)

generate_minimal_example(condition, fixed_objects = NULL)
}
\arguments{
\item{print}{print code to generate parameter set?}

\item{Design}{}

\item{condition}{row of Design dataset}

\item{fixed_objects}{list of parameters that are fixed across simulations}
}
\value{
For assumptions_mimimal_example: a design tibble with default values invisibly

For true_summary_statistics_minimal_example: the Design tibble with added column eff_size

a simulated dataset
}
\description{
Create an empty assumtions tibble for generate_minimal_example

Calculate true summary statistics value for generate_minimal_example

Generate dataset for minimal example
}
\section{Functions}{
\itemize{
\item \code{assumptions_minimal_example()}: generate default assumptions \code{tibble}

\item \code{true_summary_statistics_minimal_example()}: calculate true effect size

}}
\examples{
Design <- assumptions_minimal_example() |>
true_summary_statistics_minimal_example()

dat <- generate_minimal_example(Design[1,])
head(dat)
}
