% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_oncology.R
\name{generate_oncology}
\alias{generate_oncology}
\alias{oncology_scenario}
\alias{oncology_scenario_set_truevalues}
\title{Generate Dataset that simulates an oncology trial}
\usage{
generate_oncology(
  condition,
  fixed_objects = list(allow_switch = TRUE, logHR_assumed = NULL, ev_soll = NULL,
    allow_random_cens = TRUE)
)

oncology_scenario()

oncology_scenario_set_truevalues(Design, fixed_objects = NULL)
}
\arguments{
\item{condition}{condition row of Design dataset}

\item{fixed_objects}{fixed objects not used for now}

\item{Design}{Design data.frame for x}

\item{print}{print code to generate parameter set?}
}
\value{
For generate_oncology: A data set with n rows and the columns .....

For oncology_scenario: a design tibble with default values invisibly

For
}
\description{
Generate Dataset that simulates an oncology trial

Create an empty assumptions data.frame for generate_oncology

Calculate true summary statistics for scenarios with delayed treatment effect
}
\details{
Condition has to contain the following columns:
\itemize{
\item k              # number of post baseline visits
\item recr_interval  # recruitment interval in years (e.g. 2 means that recruitment takes 2 years)
\item max_duration   # maximal duration of the trial in years (e.g. 7 means that the last patient recruited is followed for 5 years)
\item aimed_for_n_per_required_event # This governs the recruitment rate.
\item alpha          # type I error rate for the logrank test at the end of the trial (i.e. at max_duration)
\item power          # power for the logrank test at the end of the trial (i.e. at max_duration)
\item p_trt          # proportion of patients randomized to the treatment arm
\item w              # threshold for covariate W to determine the effect of W on the hazard
\item mu_W           # list with list entries containing mean values of W at each visit for control and treatment group in k dimensional vectors
\item mu_L           # list with list entries containing mean values of L at each visit for control and treatment group in k dimensional vectors
\item Sigma_W_L      # list with matrix of dimension k x k with the covariance matrix of W and L at all visits (assuming the same covariance matrix for control and treatment group)
\item beta_prog      # list with vectors of dimension 7 with the coefficients for the progression hazard at each visit for the variables Int, X, W, W>0, L, trt, switched
\item beta_switch    # list with vectors of dimension 7 with the coefficients for the switching probability at each visit for the variables Int, X, W, W>0, L, trt, switched
\item beta_cens      # list with vectors of dimension 7 with the coefficients for the censoring hazard at each visit for the variables Int, X, W, W>0, L, trt, switched
\item beta_death     # list with vectors of dimension 8 with the coefficients for the death hazard at each visit for the variables Int, X, W, W>0, L, trt, switched, logHR_assumed
}

oncology_scenario generates a default design \code{data.frame}
for use with generate_oncology. If print is \code{TRUE} code to produce
the template is also printed for copying, pasting and editing by the user.
(This is the default when run in an interactive session.)
}
\section{Functions}{
\itemize{
\item \code{generate_oncology()}: simulates a data set with n rows.

\item \code{oncology_scenario()}: generate default design tibble

\item \code{oncology_scenario_set_truevalues()}: calculate true summary statistics for ...

}}
\examples{
Design <- oncology_scenario()
generate_oncology(Design[1, ])
Design <- oncology_scenario()[1, ]
Design
oncology_scenario_set_truevalues(oncology_scenario())
}
