% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_demediation.R,
%   R/analyse_diabetes_demediation.R
\name{summarise_diabetes_demediation}
\alias{summarise_diabetes_demediation}
\title{Summarise Output from Analyse Functions for ...}
\usage{
summarise_diabetes_demediation(name = NULL)

summarise_diabetes_demediation(name = NULL)
}
\arguments{
\item{name}{also input used to name the summarise function}
}
\value{
Returns a function with the arguments:
\itemize{
\item condition
\item results
\item fixed objects
}

that can be passed to create_summarise_function or to
SimDesign::runSimulation and that returns a \code{data.frame} with the columns
\itemize{
\item \code{Y} ...
\item ...
}

Returns a function with the arguments:
\itemize{
\item condition
\item results
\item fixed objects
}

that can be passed to create_summarise_function or to
SimDesign::runSimulation and that returns a \code{data.frame} with the columns
\itemize{
\item \code{Y} ...
\item ...
}
}
\description{
Summarise Output from Analyse Functions for ...

Summarise Output from Analyse Functions for ...
}
\section{Functions}{
\itemize{
\item \code{summarise_diabetes_demediation()}: Summarise Output from Analyse X

\item \code{summarise_diabetes_demediation()}: Summarise Output from Analyse X

}}
\examples{
summarise_diabetes_demediation("tell")
summarise_diabetes_demediation("tell")
}
