% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_MACE.R
\name{summarise_func_mace}
\alias{summarise_func_mace}
\title{Generic Summarise function}
\usage{
summarise_func_mace(condition, results, fixed_objects)
}
\arguments{
\item{condition}{SimDesign element}

\item{results}{SimDesign element}

\item{fixed_objects}{SimDesign element}
}
\value{
bias = bias on the treatment effect
mean_se = mean estimated standard error
emp_se = empirical se
mean_HR = mean estimated hazard ratio
mean_HR_CI_width = mean width of 95\% confidence interval of HR
power = power/type 1 error using a Wald test
}
\description{
Generic Summarise function
}
\examples{
internal use by SimDesign
}
