% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_vaccine.R
\name{vaccine_scenario}
\alias{vaccine_scenario}
\alias{generate_vaccine}
\title{Create an empty assumtions data.frame for generate_vaccine}
\usage{
vaccine_scenario(print = interactive())

generate_vaccine(condition, fixed_objects = NULL)
}
\arguments{
\item{print}{print code to generate parameter set?}

\item{condition}{Row of parameters dataset}

\item{fixed_objects}{other objects passed to simulation runs}
}
\value{
For vaccine_scenario: a design tibble with default values invisibly

a simulated dataset
}
\description{
Create an empty assumtions data.frame for generate_vaccine
}
\details{
vaccine_scenario generates a default design \code{data.frame}
for use with \code{generate_vaccine}. If print is \code{TRUE} code to produce
the template is also printed for copying, pasting and editing by the user.
(This is the default when run in an interactive session.)
}
\section{Functions}{
\itemize{
\item \code{vaccine_scenario()}: generate default assumptions \code{data.frame}

\item \code{generate_vaccine()}: simulate a dataset from the vaccine scenario

}}
\examples{
Design <- vaccine_scenario()
Design

Design$gamma_0 <- 0.1
Design$beta_A1 <- 0.1

generate_vaccine(Design[1,])
}
