% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_demediation.R
\name{analyse_demediation}
\alias{analyse_demediation}
\alias{summarise_demediation}
\title{Create Analyse Functions for ...}
\usage{
analyse_demediation(X)

summarise_demediation(name = NULL)
}
\arguments{
\item{X}{input can be used to pass parameters to the analyse function}

\item{name}{also input used to name the summarise function}
}
\value{
an analyse function that can be used in runSimulation

Returns a function with the arguments:
\itemize{
\item condition
\item results
\item fixed objects
}

that can be passed to create_summarise_function or to
SimDesign::runSimulation and that returns a \code{data.frame} with the columns
\itemize{
\item \code{Y} ...
\item ...
}
}
\description{
Create Analyse Functions for ...

Summarise Output from Analyse Functions for ...
}
\section{Functions}{
\itemize{
\item \code{summarise_demediation()}: Summarise Output from Analyse X

}}
\examples{
setting <- assumptions_diabetes_rescue()[1, ]
dat <- generate_diabetes_rescue(setting)
analyse_demediation()(setting, dat)

summarise_demediation("tell")
}
