% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_diabetes_demediation.R
\name{analyse_diabetes_demediation}
\alias{analyse_diabetes_demediation}
\title{Create Analyse Functions for ...}
\usage{
analyse_diabetes_demediation(X)
}
\arguments{
\item{X}{input can be used to pass parameters to the analyse function}
}
\value{
an analyse function that can be used in runSimulation
}
\description{
Create Analyse Functions for ...
}
\examples{
setting <- assumptions_diabetes_rescue()[1, ] |> true_summary_statistics_diabetes_rescue()
dat <- generate_diabetes_rescue(setting)
analyse_diabetes_demediation()(setting, dat)

}
