% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_diabetes_ipw.R
\name{analyse_diabetes_ipw}
\alias{analyse_diabetes_ipw}
\title{Analyse Dataset with the Inverse probability weighting}
\usage{
analyse_diabetes_ipw(strategy = "hypothetical")
}
\arguments{
\item{strategy}{the estimand targeted by the CI method, default: "hypothetical". Allowed also: "treatment_policy".}
}
\value{
an analyse function that can be used in runSimulation that returns a list with the elements
\itemize{
\item \code{coef} coefficient for \code{trt}
\item \code{sd} standard deviation for coef
\item \code{p} p-value for coef
\item \code{ci_lower} lower bound of confidence interval for coef
\item \code{ci_upper} upper bound of confidence interval for coef
}
}
\description{
Analyse Dataset with the Inverse probability weighting
}
\examples{
Design <- diabetes_scenario() |>
  diabetes_scenario_set_truevalues()

condition <- Design[1, ]

dat <- generate_diabetes(condition)

analyse_diabetes_ipw(strategy = "treatment_policy")(condition, dat)
analyse_diabetes_ipw(strategy = "hypothetical")(condition, dat)

}
