% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_diabetes_ipw.R
\name{analyse_diabetes_ipw}
\alias{analyse_diabetes_ipw}
\title{Analyse Dataset with the Inverse probability weighting}
\usage{
analyse_diabetes_ipw(estimand = "hyp")
}
\arguments{
\item{estimand}{the estimand targeted by the CI method, default: "hyp" (hypothetical). Allowed also: "tp" (treatment policy).}
}
\value{
an analyse function that can be used in runSimulation that returns a list with the elements
\itemize{
\item \code{coef} coefficient for \code{trt}
\item \code{sd} standard deviation for coef
}
}
\description{
Analyse Dataset with the Inverse probability weighting
}
\examples{
Design <- assumptions_diabetes_rescue() |>
  true_summary_statistics_diabetes_rescue()

condition <- Design[1, ]

dat <- generate_diabetes_rescue(condition)

analyse_diabetes_ipw_tp <- analyse_diabetes_ipw(estimand = "tp")
analyse_diabetes_ipw_tp(condition, dat)

analyse_diabetes_ipw_hyp <- analyse_diabetes_ipw(estimand = "hyp")
analyse_diabetes_ipw_hyp(condition, dat)
}
