% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_diabetes_mi.R
\name{analyse_diabetes_mi}
\alias{analyse_diabetes_mi}
\title{Analyse data set with Multiple imputation}
\usage{
analyse_diabetes_mi(
  strategy = c("hypothetical", "treatment_policy"),
  m = 10,
  maxit = 20,
  ci_level = 0.95,
  seed = 123
)
}
\arguments{
\item{strategy}{Either \code{"hypothetical"} or \code{"treatment_policy"}.}

\item{m}{Number of imputations (used for hypothetical only).}

\item{maxit}{Maximum number of MICE iterations.}

\item{ci_level}{confidence level for the CI (default 0.95)}

\item{seed}{Random seed for imputation.}
}
\value{
A function that, when called with \code{condition} and \code{dat}, returns a list with:
\itemize{
\item \code{coef} coefficient for \code{trt}
\item \code{p} p-value for coef
\item \code{ci_lower} lower bound of \code{ci-level}\% confidence interval for coef
\item \code{ci_upper} upper bound of \code{ci-level}\% confidence interval for coef
}
}
\description{
Creates an analysis function for the diabetes rescue simulation.
Depending on the chosen estimand, either a hypothetical strategy
with multiple imputation after rescue initiation is applied,
or a treatment policy strategy using observed data only.
}
\examples{
Design <- diabetes_scenario()[1, ] |>
  diabetes_scenario_set_truevalues()

dat <- generate_diabetes(Design)

analyse_diabetes_mi(strategy = "treatment_policy")(Design, dat)
analyse_diabetes_mi(strategy = "hypothetical")(Design, dat)

}
