% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_diabetes_mmrm.R
\name{analyse_diabetes_mmrm}
\alias{analyse_diabetes_mmrm}
\title{Analyse diabetes endpoint using MMRM}
\usage{
analyse_diabetes_mmrm(
  ci_level = 0.95,
  strategy = c("treatment_policy", "hypothetical")
)
}
\arguments{
\item{ci_level}{confidence level for the CI (default 0.95)}

\item{strategy}{strategy to handle rescue medication: either "treatment_policy" (use all observed data) or "hypothetical" (set post-rescue values to missing and ignore them in the analysis)}
}
\value{
an analyse function that returns a list with
\itemize{
\item \code{p} p-value for treatment effect at final visit
\item \code{coef} estimated treatment effect at final visit
\item \code{ci_lower} lower confidence limit
\item \code{ci_upper} upper confidence limit
}
}
\description{
Analyse diabetes endpoint using MMRM
}
\examples{
setting <- diabetes_scenario()[1, ] |> diabetes_scenario_set_truevalues()
dat <- generate_diabetes(setting)
analyse_diabetes_mmrm()(setting, dat)

}
