% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_diabetes_mmrm.R
\name{analyse_diabetes_mmrm}
\alias{analyse_diabetes_mmrm}
\title{Analyse diabetes endpoint using MMRM}
\usage{
analyse_diabetes_mmrm(
  ci_level = 0.95,
  strategy = c("treatment_policy", "hypothetical")
)
}
\arguments{
\item{ci_level}{confidence level for the CI (default 0.95)}

\item{strategy}{strategy to handle rescue medication: either "treatment_policy" (use all observed data) or "hypothetical" (set post-rescue values to missing and ignore them in the analysis)}
}
\value{
A function that, when called with \code{condition} and \code{dat}, returns a list with:
\itemize{
\item \code{coef} coefficient for \code{trt}
\item \code{p} p-value for coef
\item \code{ci_lower} lower bound of 95\% confidence interval for coef
\item \code{ci_upper} upper bound of 95\% confidence interval for coef
}
}
\description{
Analyse diabetes endpoint using MMRM
}
\details{
This function implements a mixed model for repeated measures (MMRM) to estimate the treatment effect on the change in HbA1c at the final visit in a diabetes trial, accounting for rescue medication use.
The function first reshapes the data to a long format, where each row corresponds to a visit for a patient.
Depending on the chosen strategy, it either uses all observed data (treatment policy) or sets post-rescue values to missing (hypothetical).
The MMRM is then fitted with an unstructured covariance matrix for the repeated measures, adjusting for baseline HbA1c, age, and treatment, nested within visits.
Finally the treatment effect at the final visit is extracted along with its standard error and confidence interval.
}
\examples{
Design <- diabetes_scenario()[1, ] |>
  diabetes_scenario_set_truevalues()

dat <- generate_diabetes(Design)

analyse_diabetes_mmrm()(Design, dat)

}
