% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_diabetes_mmrm.R,
%   R/analyse_diabetes_rescue.R
\name{analyse_diabetes_rescue_mmrm}
\alias{analyse_diabetes_rescue_mmrm}
\title{Analyse dataset from diabetes rescue scenario using MMRM}
\usage{
analyse_diabetes_rescue_mmrm(ci_level = 0.95)

analyse_diabetes_rescue_mmrm(ci_level = 0.95)
}
\arguments{
\item{ci_level}{the confidence level for the CIs (defaults to 0.95)}
}
\value{
an analyse function that returns a list with
\itemize{
\item \code{p} p-value for treatment effect at final visit
\item \code{coef} estimated treatment effect at final visit
\item \code{ci_lower} lower confidence limit
\item \code{ci_upper} upper confidence limit
}

an analyse function that returns a list with the elements
\itemize{
\item \code{p} p-value for the treatment effect at the final visit
\item \code{coef} estimated treatment effect at the final visit
\item \code{ci_lower} the lower CI limit
\item \code{ci_upper} the upper CI limit
}
}
\description{
Analyse dataset from diabetes rescue scenario using MMRM

Analyse dataset from diabetes rescue scenario using MMRM
}
\examples{
Design <- assumptions_diabetes_rescue()
dat <- generate_diabetes_rescue(Design[1, ])
analyse_mmrm <- analyse_diabetes_rescue_mmrm()
analyse_mmrm(Design[1, ], dat)
Design <- assumptions_diabetes_rescue() |>
  true_summary_statistics_diabetes_rescue()

condition <- Design[1, ]

dat <- generate_diabetes_rescue(condition)

analyse_mmrm <- analyse_diabetes_rescue_mmrm(ci_level = 0.95)
analyse_mmrm(condition, dat)
}
