% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_oncology_gformula.R
\name{analyse_oncology_gformula}
\alias{analyse_oncology_gformula}
\title{Create Analyse Functions for g formula}
\usage{
analyse_oncology_gformula(B = 20, reps = 1)
}
\arguments{
\item{B}{Number of bootstrap samples}

\item{reps}{Number of repetitions for the g-formula simulation (per bootstrap sample)}
}
\value{
an analyse function that can be used in runSimulation
}
\description{
Create Analyse Functions for g formula
}
\examples{
\donttest{
setting <- oncology_scenario()[1, ]

dat <- generate_oncology(setting)

analyse_oncology_gformula()(setting, dat)
}
}
