% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_oncology_TSE.R
\name{analyse_oncology_TSE}
\alias{analyse_oncology_TSE}
\title{Create Analyse Functions for}
\usage{
analyse_oncology_TSE(recensor = TRUE, alpha = 0.05, B = 100)
}
\arguments{
\item{recensor}{logical, whether to recensor the data after adjusting for treatment switching. Default is TRUE.}

\item{alpha}{numeric, significance level for confidence intervals. Default is 0.05.}

\item{B}{integer, number of bootstrap samples for estimating standard errors. Default is 100.}
}
\value{
an analyse function that can be used in runSimulation
}
\description{
Create Analyse Functions for
}
\examples{
setting <- oncology_scenario()[1, ]
dat <- generate_oncology(setting)
analyse_oncology_TSE()(setting, dat)
}
