% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_vaccine_ivreg.R
\name{analyse_vaccine_ivreg}
\alias{analyse_vaccine_ivreg}
\title{Analyse dataset from vaccine scenario using instrumental variable regression}
\usage{
analyse_vaccine_ivreg(ci_level = 0.95, VE_margin = 0.3)
}
\arguments{
\item{ci_level}{the confidence level for the CIs (defaults to 0.95)}

\item{VE_margin}{vaccine efficacy margin for the super-superiority test}
}
\value{
an analyse function that returns a list with the elements
\itemize{
\item \code{p} the p-value of the exact binomial test
\item \code{VE} the point estimate for the vaccine efficacy
\item \code{VE_lower} the lower CI limit
\item \code{VE_upper} the upper CI limit
}
}
\description{
Analyse dataset from vaccine scenario using instrumental variable regression
}
\examples{
Design <- vaccine_scenario() |>
  vaccine_scenario_set_gamma_0() |>
  vaccine_scenario_set_true_eff() |>
  vaccine_scenario_set_samplesize()
Design

dat <- generate_vaccine(Design[1,])
my_analyse <- analyse_vaccine_ivreg(ci_level=0.95)
my_analyse(Design[1, ], dat)
}
