% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_vaccine_pp.R
\name{analyse_vaccine_pp}
\alias{analyse_vaccine_pp}
\title{Analyse dataset from vaccine scenario, per-protocol analysis}
\usage{
analyse_vaccine_pp(ci_level = 0.95)
}
\arguments{
\item{ci_level}{the confidence level for the CIs (defaults to 0.95)}
}
\value{
an analyse function that returns a list with the elements
\itemize{
\item \code{p} the p-value of the exact binomial test
\item \code{coef} the point estimate
\item \code{ci_lower} the lower CI limit
\item \code{ci_upper} the upper CI limit
}
}
\description{
Analyse dataset from vaccine scenario, per-protocol analysis
}
\examples{
Design <- vaccine_scenario() |>
  vaccine_scenario_set_gamma_0() |>
  vaccine_scenario_set_true_eff() |>
  vaccine_scenario_set_samplesize()
Design

Design$beta_A1 <- 0.1

dat <- generate_vaccine(Design[1,])
my_analyse <- analyse_vaccine_pp(ci_level=0.95)
my_analyse(Design[1, ], dat)
}
