% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_vaccine_ps.R
\name{analyse_vaccine_ps}
\alias{analyse_vaccine_ps}
\title{Analyse dataset from vaccine scenario using principal score weighting}
\usage{
analyse_vaccine_ps(
  ci_level = 0.95,
  VE_margin = 0.3,
  covariates_in_outcomes_model = TRUE,
  V_unobserved = FALSE,
  W_unobserved = FALSE,
  W_interaction = FALSE
)
}
\arguments{
\item{ci_level}{the confidence level for the CIs (defaults to 0.95)}

\item{VE_margin}{vaccine efficacy margin for the super-superiority test}

\item{covariates_in_outcomes_model}{should the covariates to estimate the principal score also be included in the outcomes model}

\item{V_unobserved}{consider covariate V unobserved (don't use it in analysis)}

\item{W_unobserved}{consider covariate W unobserved (don't use it in analysis)}

\item{W_interaction}{add treatment x W interaction in outcomes model}
}
\value{
an analyse function that returns a list with the elements
\itemize{
\item \code{p_model}  p-value of  super-superiority test, model based vcov
\item \code{VE}  point estimate for  vaccine efficacy, model based vcov
\item \code{VE_lower}  lower CI limit for vaccine efficacy, model based vcov
\item \code{VE_upper}  upper CI limit for vaccine efficacy, model based vcov
\item \code{OR}  point estimate for  odds-ratio for infection, model based vcov
\item \code{OR_lower}  lower CI limit for  odds-ratio for infection, model based vcov
\item \code{OR_upper}  upper CI limit for  odds-ratio for infection, model based vcov
\item \code{OR_sandwich}  point estimate for  odds-ratio for infection, robust vcov
\item \code{OR_lower_sandwich}  lower CI limit for  odds-ratio for infection, robust vcov
\item \code{OR_upper_sandwich}  upper CI limit for  odds-ratio for infection, robust vcov
\item \code{p}  p-value of  super-superiority test, robust vcov
\item \code{VE_sandwich}  point estimate for  vaccine efficacy, robust vcov
\item \code{VE_lower_sandwich}  lower CI limit for vaccine efficacy, robust vcov
\item \code{VE_upper_sandwich}  upper CI limit for vaccine efficacy, robust vcov
\item \code{N_pat} number of patients in the dataset
\item \code{N_evt} number of observed events in the dataset
}
}
\description{
Analyse dataset from vaccine scenario using principal score weighting
}
\examples{
Design <- vaccine_scenario() |>
  vaccine_scenario_set_beta_A1_relative() |>
  vaccine_scenario_set_gamma_0() |>
  vaccine_scenario_set_true_eff() |>
  vaccine_scenario_set_samplesize()

dat <- generate_vaccine(Design[3,])
my_analyse <- analyse_vaccine_ps(ci_level=0.95)
my_analyse(Design[3, ], dat)
}
