% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.RCT.Sim_package.R
\docType{package}
\name{CI.RCT.Sim}
\alias{CI.RCT.Sim-package}
\alias{CI.RCT.Sim}
\title{CI.RCT.Sim}
\description{
CI.RCT.Sim: simulate scenarios in which causal inference methods are applied in randomized controlled trials
}
\details{
Package to Facilitate a Simulation Study on Causal Inference Methods in RCTs
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ci-rct-sim.github.io/CI.RCT.Sim/}
}

}
\author{
\strong{Maintainer}: Tobias Fellinger \email{tobias.fellinger@ages.at} (\href{https://orcid.org/0000-0001-9474-2731}{ORCID})

Authors:
\itemize{
  \item Tobias Fellinger \email{tobias.fellinger@ages.at} (\href{https://orcid.org/0000-0001-9474-2731}{ORCID})
}

}
\keyword{internal}
