% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_diabetes_rescue.R
\name{generate_diabetes}
\alias{generate_diabetes}
\alias{diabetes_scenario_set_truevalues}
\title{Generate Dataset that simulates a diabetes trial in which rescue medication
can be introduced}
\usage{
generate_diabetes(condition, fixed_objects = NULL)

diabetes_scenario_set_truevalues(
  Design,
  cutoff_stats = 10,
  fixed_objects = NULL
)
}
\arguments{
\item{condition}{condition row of Design dataset}

\item{fixed_objects}{fixed objects not used for now}

\item{Design}{Design data.frame for x}

\item{cutoff_stats}{Cutoff time for rmst and average hazard ratios}
}
\value{
For generate_diabetes: A data set with n rows and the columns id, trt
(1=treatment, 0=control), age, y0, y1, ..., yk
(the repeated measurements of the outcome), R1, ..., Rk-1 (indicators for
rescue medication at each visit except baseline and the last visit),
and rescue_start (the visit at which rescue medication was started,
NA if no rescue medication was received).

For true_summary_statistics_x: the design data.frame
passed as argument with the additional columns:
\itemize{
\item \code{rmst_trt} rmst in the treatment group
\item \code{median_surv_trt} median survival in the treatment group
\item \code{rmst_ctrl} rmst in the control group
\item \code{median_surv_ctrl} median survial in the control group
\item \code{gAHR} geometric average hazard ratio
\item \code{AHR} average hazard ratio
}
}
\description{
Generate Dataset that simulates a diabetes trial in which rescue medication
can be introduced

Calculate true summary statistics for scenarios with delayed treatment effect
}
\details{
Condition has to contain the following columns:
\itemize{
\item k
\item mean_age
\item sd_age
\item b_age
\item mean_bl
\item sd_bl
\item rho
\item delta
\item lambda
\item delta_resc
\item lambda_resc
\item resc_0
\item resc_y
\item resc_age
\item miss_0
\item miss_y
\item miss_age
\item miss_resc
}
}
\section{Functions}{
\itemize{
\item \code{generate_diabetes()}: simulates a data set with n rows.

\item \code{diabetes_scenario_set_truevalues()}: calculate true summary statistics for ...

}}
\examples{
Design <- diabetes_scenario()
generate_diabetes(Design[1, ])
diabetes_scenario_set_truevalues(diabetes_scenario())
}
