% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MACE.R
\name{assumptions_mace}
\alias{assumptions_mace}
\alias{generate_mace}
\alias{true_trt_mace}
\title{Generate Data for MACE Scenario}
\usage{
assumptions_mace()

generate_mace(condition, fixed_objects)

true_trt_mace(Design)
}
\arguments{
\item{condition}{row of Design dataset}

\item{fixed_objects}{list of parameters that are fixed across simulations}

\item{Design}{Design dataset}
}
\value{
A data.frame with parameter Values for scenarios

a simulated dataset

Design data set with additional column \code{true_trt}
}
\description{
Generate Data for MACE Scenario
}
\section{Functions}{
\itemize{
\item \code{assumptions_mace()}: generate default assumptions \code{tibble}

\item \code{true_trt_mace()}: Calculate true treatment effect for MACE scenario

}}
\examples{
Design <- assumptions_mace() |>
dat <- generate_mace(Design[1,])
head(dat)
}
