% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_minimal_example.R
\name{generate_minimal_example}
\alias{generate_minimal_example}
\alias{true_summary_statistics_minimal_example}
\alias{assumptions_minimal_example}
\title{Generate dataset for minimal example}
\usage{
generate_minimal_example(condition, fixed_objects = NULL)

true_summary_statistics_minimal_example(Design)

assumptions_minimal_example(print = interactive())
}
\arguments{
\item{condition}{row of Design dataset}

\item{fixed_objects}{list of parameters that are fixed across simulations}

\item{Design}{}

\item{print}{print code to generate parameter set?}
}
\value{
a simulated dataset

For true_summary_statistics_minimal_example: the Design tibble with added column eff_size

For assumptions_mimimal_example: a design tibble with default values invisibly
}
\description{
Generate dataset for minimal example

Calculate true summary statistics value for generate_minimal_example

Create an empty assumtions tibble for generate_minimal_example
}
\section{Functions}{
\itemize{
\item \code{true_summary_statistics_minimal_example()}: calculate true effect size

\item \code{assumptions_minimal_example()}: generate default assumptions \code{tibble}

}}
\examples{
Design <- assumptions_minimal_example() |>
true_summary_statistics_minimal_example()

dat <- generate_minimal_example(Design[1,])
head(dat)
}
