% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{create_summarise_function}
\alias{summarise_estimator}
\alias{summarise_test}
\alias{upsert_merge}
\alias{merge_additional_results}
\alias{rename_results_column}
\alias{rename_results_column_pattern}
\alias{results_pivot_longer}
\alias{combined_plot}
\alias{labs_from_labels}
\alias{wrap_all_in_trycatch}
\alias{wrap_all_in_preserve_seed}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{SimNPH}{\code{\link[SimNPH:combined_plot]{combined_plot()}}, \code{\link[SimNPH:create_summarise_function]{create_summarise_function()}}, \code{\link[SimNPH:labs_from_labels]{labs_from_labels()}}, \code{\link[SimNPH:merge_additional_results]{merge_additional_results()}}, \code{\link[SimNPH:rename_results_column]{rename_results_column()}}, \code{\link[SimNPH:rename_results_column_pattern]{rename_results_column_pattern()}}, \code{\link[SimNPH:results_pivot_longer]{results_pivot_longer()}}, \code{\link[SimNPH:summarise_estimator]{summarise_estimator()}}, \code{\link[SimNPH:summarise_test]{summarise_test()}}, \code{\link[SimNPH:upsert_merge]{upsert_merge()}}, \code{\link[SimNPH:wrap_all_in_preserve_seed]{wrap_all_in_preserve_seed()}}, \code{\link[SimNPH:wrap_all_in_trycatch]{wrap_all_in_trycatch()}}}
}}

