% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_vaccine.R
\name{vaccine_scenario}
\alias{vaccine_scenario}
\alias{generate_vaccine}
\alias{vaccine_scenario_set_gamma_0}
\alias{vaccine_scenario_set_true_eff}
\alias{vaccine_scenario_set_samplesize}
\title{Create an empty assumptions data.frame for generate_vaccine}
\usage{
vaccine_scenario(print = interactive())

generate_vaccine(condition, fixed_objects = list(include_unobserved = FALSE))

vaccine_scenario_set_gamma_0(Design, r = 1)

vaccine_scenario_set_true_eff(Design)

vaccine_scenario_set_samplesize(
  Design,
  alpha = 0.025,
  CSE = 0.3,
  VE_H1 = 0.7,
  power = 0.8,
  r = 1
)
}
\arguments{
\item{print}{print code to generate parameter set?}

\item{condition}{Row of parameters dataset}

\item{fixed_objects}{other objects passed to simulation runs}

\item{Design}{Design dataset as returned by \code{vaccine_scenario}}

\item{r}{for setting gamm_0 and sample size calculation: allocation ratio}

\item{alpha}{for sample size calculation: one-sided alpha level}

\item{CSE}{for sample size calculation: super-superiority margin on the VE scale}

\item{VE_H1}{vaccine efficacy assumed for sample size planning}

\item{power}{for sample size calculation: target power}
}
\value{
For vaccine_scenario: a design tibble with default values invisibly

a simulated dataset

a Design dataset with the added column gamma_0

a Design dataset with the added column rr_ps

a Design dataset with the added columns n_trt and n_ctrl
}
\description{
Create an empty assumptions data.frame for generate_vaccine
}
\details{
vaccine_scenario generates a default design \code{data.frame}
for use with \code{generate_vaccine}. If print is \code{TRUE} code to produce
the template is also printed for copying, pasting and editing by the user.
(This is the default when run in an interactive session.)
}
\section{Functions}{
\itemize{
\item \code{vaccine_scenario()}: generate default assumptions \code{data.frame}

\item \code{generate_vaccine()}: simulate a dataset from the vaccine scenario

\item \code{vaccine_scenario_set_gamma_0()}: vaccine_scenario_set_gamma_0 calculate gamma_0 from other parameters

\item \code{vaccine_scenario_set_true_eff()}: vaccine_scenario_set_true_eff calculate relative risk in the principal stratum of always compliers

\item \code{vaccine_scenario_set_samplesize()}: vaccine_scenario_set_samplesize calculate sample size as if the true effect was correctly guessed

}}
\examples{
Design <- vaccine_scenario() |>
  vaccine_scenario_set_gamma_0() |>
  vaccine_scenario_set_true_eff() |>
  vaccine_scenario_set_samplesize()
Design

generate_vaccine(Design[1,])
}
